requirements.txt
websocket-client
pandas
numpy
ta
joblib
lightgbm
